Initial Import
Installs base icinga2 packages, a subset of features and icingaweb2 in a
modular fashion.  This facilitates provisioning of icinga2 alone as a
satellite or active checker, and also adding IDO and web interfaces at
any specific point in the hierarchy.  Initial support for endpoints and
zone orchestration if you require explicit control.