Multiple Target Support For Zones
Allow the zone and endpoint types to operate locally on zones.conf or populate
repository.d if in repository mode.  This keeps configuration managed in the
same way as icinga2 when using the CLI commands.  Additionally the zone.conf.erb
script has been updated to cast to an Array as puppet will convert arrays with a
single element to a