Add Host Management Support
Performs rudimentary host configuration locally.  In repository mode for
collected resources if the host is directly connected to a satellite or
master node then cluster-zone checks are performed.  If the host is not
directly connected the check command is a dummy and the satellite zone
performing the check is selected.