Improve Service Handling
When services are collected on a satellite or master we want to override the
check command selectively based on tags.  As such roll back the inteligence
from the service template and make the check name explicit as overriding the
check command would result in a bunch of commands being called dummy on the
satellite.