Dynamic Display Names
Prototype specifying a display_name in apply_sservice_for as either 'service'
or '"service " + key'.  The template dynamically decides whether to quote the
variable based on whether the string contains any quotes itself.  This can be
rolled out to all parameters in future via a library function to generate the
correct configuration strings