Improve Template Generation
Creates a utility function that can be used in templates to generate
correct syntax.  It translates native booleans, integers and floats
to unquoted parameters.  Regular strings are parsed to determine whether
they are time periods, references to hosts, services, users or key/value
pairs in 'apply service for' objects, or preformatted strings.  Again
these are unquoted, anything else is assumed a string expression and
quoted.  Hashes are also supported as is simple detection of whether
keys need to be quoted as they contain illegal characters e.g. -b in
check command arguments