Icingaweb2 Fixes
Handles some updates to icingaweb2 that drop privs to the icingaweb2 user.
Actually installs the IDO schema for icinga web, installs the checker
feature so we can test checks are working via the API later on.  Adds
some host and service templates.  Updates the test VM to define the
FQDN in /etc/hosts so this is picked up by 'hostname --fqdn' required
by the HostName constant.